pkg/lexer/cuda_test.go
package lexer_test

import (
	"os"
	"testing"

	"github.com/wakatime/wakatime-cli/pkg/lexer"

	"github.com/stretchr/testify/assert"
)

func TestCUDA_AnalyseText(t *testing.T) {
	tests := map[string]struct {
		Filepath string
		Expected float32
	}{
		"include": {
			Filepath: "testdata/cuda_include.cu",
			Expected: 0.1,
		},
		"ifdef": {
			Filepath: "testdata/cuda_ifdef.cu",
			Expected: 0.1,
		},
		"ifndef": {
			Filepath: "testdata/cuda_ifndef.cu",
			Expected: 0.1,
		},
	}

	for name, test := range tests {
		t.Run(name, func(t *testing.T) {
			data, err := os.ReadFile(test.Filepath)
			assert.NoError(t, err)

			l := lexer.CUDA{}.Lexer()

			assert.Equal(t, test.Expected, l.AnalyseText(string(data)))
		})
	}
}
